requirements.txt
torch
torchvision
gradio
gradio-imageslider
pillow
numpy
opencv-python
basicsr
realesrgan
gfpgan